Add files via upload
Client: Social Buzz.

Social Buzz has reached huge scale in recent years to become recognized as a global unicorn company. It is a social media and content creation company. Over the past 5 years, Social Buzz has reached over 500 million active users each month. Their platform receives over 100000 posts per day which amounts to 36 500 000 posts every year, of which, this is all unstructured data making it very hard to make sense of.

They are expecting the following:

1. An audit of their big data practice.
2. Recommendations for a successful IPO.
3. An analysis of their content categories that highlights the top 5 categories with the largest aggregate popularity.

Focusing on the last point that I mentioned there, this is what the Data Analytics team has been specifically focused on. Clearly with such grand scale, this comes with a lot of data and with such vast amounts of data comes challenges.

But how to capitalize on it when there is so much?

It's not just all about harvesting as much content as possible... The real value is in understanding and crunching this content to gain a deeper understanding of their audience and to therefore provide a more personalized and enjoyable experience.
And this is where the data analytics expertise comes in.

So, how did I tackle this problem?
Well I approached it in 5 steps:
1. Data understanding - the key to success on any data project is to understand the data in detail. So I took the time to understand the data model and domain of their business.
2. Data cleaning - after understanding their business, I then cleaned the available datasets and thought about what an ideal dataset should look like for this problem.
3. Data modeling - After ensuring the data was clean for analysis, I needed to process and model this data into a dataset that can precisely answer the business questions and produce the results needed.
4. Data analysis - With this new dataset, I used my analytical skills to uncover insights from this dataset and to produce visualizations to describe the insights.
5. And finally I used these insights to unlock business decisions.